Small improvement to hash computation.

+++ example/hash-set/src/main.c
@@ -9,7 +9,7 @@
 #include <time.h>
 #include "input.h"
 
-#define SEED ((uint64_t)time(NULL))
+#define SEED ((uint32_t)time(NULL))
 
 /* ========================================================================= */
 /* MAIN                                                                      */
